Example for loading images using Ion. (Issue #4)
Provided a short example on loading web images using Ion for issue #4.

Unfortunately, I wasn't able to get a custom Transformation to work with
Ion. The only way I know of successfully loading images with Ion would
be to use the Ion.with(Context).load(url).asBitmap() API with a callback
for image handling.

Nevertheless, I'll continue searching for a proper solution to this.